--- client/components/AskForSuggestion/AskForSuggestion.ts
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const AskForSuggestion = () => {
-
-}
-
-export default AskForSuggestion;
