--- client/components/Homepage/Homepage.ts
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Homepage = () => {
-
-}
-
-export default Homepage;
