--- client/components/Login/Login.ts
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Login = () => {
-  
-}
-
-export default Login;
