--- client/components/Recommend/Recommend.ts
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Recommend = () => {
-
-}
-
-export default Recommend;
