--- client/components/Signup/Signup.ts
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Signup = () => {
-  
-}
-
-export default Signup;
