--- client/components/UserProfile/UserProfile.ts
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const UserProfile = () => {
-  
-}
-
-export default UserProfile;
