+++ client/components/pages/AskForSuggestion.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const AskForSuggestion = () => {
+  return(
+    <div>hello</div>
+  )
+};
+
+export default AskForSuggestion;
