+++ client/components/pages/Homepage.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const Homepage = () => {};
+
+export default Homepage;
