+++ client/components/pages/Recommend.tsx
@@ -1,0 +1,4 @@
+import React from 'react';
+const Recommend = () => {};
+
+export default Recommend;
