+++ client/components/pages/UserProfile.tsx
@@ -1,0 +1,4 @@
+import React from 'react';
+const UserProfile = () => {};
+
+export default UserProfile;
