+++ client/components/styles/NavbarStyles.scss
